[Gold V] Title: 주사위, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Gold/1041. 주사위/주사위.cc
@@ -8,7 +8,7 @@ long long A[6];
 long long a, b, c;
 
 long long step1() {
-    long long val = min(min(a, b), c);
+    int val = min(min(a, b), c);
 
     return val * ((N - 1) * (N - 2) * 4 + (N - 2) * (N - 2));
 }
